PUBDEV-5378: Extended Frame's API with uniqueValuesBy(column) method. Moved MRTasks into a dedicated package.
